fix: refactor rags
