+++ packages/@aws-cdk/toolkit-lib/lib/api/tags/index.ts
@@ -1,0 +1,1 @@
+export * from './tags';
