+++ packages/@aws-cdk/toolkit-lib/lib/api/tags/private/util.ts
@@ -1,14 +1,9 @@
 import type * as cxapi from '@aws-cdk/cx-api';
+import type { Tag } from '../tags';
 
 /**
  * @returns an array with the tags available in the stack metadata.
- * @internal
  */
 export function tagsForStack(stack: cxapi.CloudFormationStackArtifact): Tag[] {
   return Object.entries(stack.tags).map(([Key, Value]) => ({ Key, Value }));
-}
-
-export interface Tag {
-  readonly Key: string;
-  readonly Value: string;
 }
