+++ packages/@aws-cdk/toolkit-lib/lib/api/tags/tags.ts
@@ -1,0 +1,4 @@
+export interface Tag {
+  readonly Key: string;
+  readonly Value: string;
+}
