+++ packages/aws-cdk/lib/api-private.ts
@@ -2,6 +2,7 @@
 export { deployStack, DeployStackOptions as DeployStackApiOptions } from '../../@aws-cdk/toolkit-lib/lib/api/deployments/deploy-stack';
 export * as cfnApi from '../../@aws-cdk/toolkit-lib/lib/api/deployments/cfn-api';
 export * from '../../@aws-cdk/toolkit-lib/lib/api/io/private';
+export * from '../../@aws-cdk/toolkit-lib/lib/api/tags/private';
 export * from '../../@aws-cdk/toolkit-lib/lib/private/activity-printer';
 export * from '../../@aws-cdk/toolkit-lib/lib/api/cloud-assembly/private/borrowed-assembly';
